% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_clinical_detection_prob.R
\name{calculate_clinical_detection_prob}
\alias{calculate_clinical_detection_prob}
\title{Calculate probability of clinical detection}
\usage{
calculate_clinical_detection_prob(
  nts,
  det,
  delay_onset,
  delay_seek,
  delay_diag,
  prob_symptomatic,
  prob_seek_healthcare,
  prob_diagnosis
)
}
\arguments{
\item{nts}{Output of [generate_number_shedding_time_series()]}

\item{det}{Output of [calculate_wastewater_ttd()]}

\item{delay_onset}{Numeric vector giving the probability distribution of
the delay from infection to symptom onset.}

\item{delay_seek}{Numeric vector giving the probability distribution of
the delay from symptom onset to seeking healthcare.}

\item{delay_diag}{Numeric vector giving the probability distribution of
the delay from seeking healthcare to diagnosis.}

\item{prob_symptomatic}{Numeric probability that an infected individual
develops symptoms.}

\item{prob_seek_healthcare}{Probability that a symptomatic
individual seeks healthcare.}

\item{prob_diagnosis}{Probability that an individual seeking
healthcare is diagnosed.}
}
\value{
The object supplied to `det$sampled_data`,
  with an additional `prob_detection_clinical` element containing the
  probability of clinical detection for each day.
}
\description{
Calculates the probability of clinical detection over time based on
wastewater-derived infections and distributions describing delays from
infection to symptom onset, healthcare seeking, and diagnosis.
}
\details{
The delay distributions are normalised before being convolved to obtain
the distribution of the total delay from infection to diagnosis. The
resulting probabilities are then used to calculate the probability of
observing at least one clinical detection for each day.
}
