% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wastewater_surveillance.R
\name{calculate_wastewater_ttd}
\alias{calculate_wastewater_ttd}
\title{Calculate wastewater time to detection}
\usage{
calculate_wastewater_ttd(
  wastewater_number_shedding_time_series,
  sampling_frequency,
  sampling_method,
  detection_approach,
  detection_params
)
}
\arguments{
\item{wastewater_number_shedding_time_series}{Output from generate_number_shedding_time_series}

\item{sampling_frequency}{The frequency of wastewate sampling, with 1 being daily, 7 being weekly, 14 being fortnightly}

\item{sampling_method}{The method used, either "autosampler", "grab" or "moore_swab"}

\item{detection_approach}{Approach to translating shedding into probability of detection, either "threshold", "logistic_curve" or "per_person_probability"}

\item{detection_params}{A named list. If detection_approach == "threshold", must contain "threshold_limits" and "population". If detection_approach == "logistic_curve", must contain "logistic_beta_0", "logistic_beta_1", "seed", "limit_of_detection", "population". If detection_approach == "per_person_probability", must contain "per_infection_probability_detection". Additionally, if "sampling_method" == "moore_swab", "detection_params must also contain "duration" in days (>=1).}
}
\description{
Calculates the time to detection of wastewater shedding using the
specified sampling method and detection approach.
}
