% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_spillover.R
\name{circular_dist}
\alias{circular_dist}
\title{Generate spillover cases}
\usage{
circular_dist(t, tmax, P = 365)
}
\arguments{
\item{t}{The day (base = 1, min = 1, max = Inf)}

\item{tmax}{The day of the year that the variable peaks (min = 1, max = 365)}

\item{P}{The period of seasonality in days (default = 365)}

\item{b}{The baseline of the variable outside of the peak}

\item{d}{The seasonal forcing (amplitude increase rel to baseline, default = 0)}

\item{sigma}{The spread of the peak (sd of underlying normal distribution)}
}
\description{
These functions define the daily probability of spillover, which can be time varying.
They then simulate the occurrence of spillovers over a time period of interest
}
