% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_branching_process.R
\name{extract_outbreak_characteristics_time}
\alias{extract_outbreak_characteristics_time}
\title{Extracts relevant information about the outbreak occurring nearest to the}
\usage{
extract_outbreak_characteristics_time(
  outbreak_info,
  outbreak_id,
  id_type = "time"
)
}
\arguments{
\item{outbreak_info}{Output from generate_outbreak_size}

\item{outbreak_id}{An identifier for the outbreak you want to extract info for - either time or outbreak number}

\item{id_type}{Either time or outbreak number}
}
\description{
Extracts relevant information about the outbreak occurring nearest to the
}
\seealso{
Other post-processing: 
\code{\link{generate_healthcare_seeking_time_series}()},
\code{\link{generate_infections_time_series}()},
\code{\link{generate_number_shedding_time_series}()},
\code{\link{generate_outbreak_size}()},
\code{\link{generate_seropositivity_timeseries}()},
\code{\link{generate_symptom_onset_time_series}()}
}
\concept{post-processing}
