% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_branching_process.R
\name{generate_hospitalisation_time_series}
\alias{generate_hospitalisation_time_series}
\title{Convert stochastic realisation of branching process into hospitalisation time-series}
\usage{
generate_hospitalisation_time_series(branching_process_output, population)
}
\description{
This function converts a branching process output into time-series of hospitalisations
and hospital occupancy.
}
\seealso{
Other post-processing: 
\code{\link{generate_infections_time_series}()},
\code{\link{generate_number_shedding_time_series}()},
\code{\link{generate_symptom_onset_time_series}()}
}
\concept{post-processing}
