% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_branching_process.R
\name{generate_infections_time_series}
\alias{generate_infections_time_series}
\title{Convert stochastic realisation of branching process into number of new infections time-series}
\usage{
generate_infections_time_series(branching_process_output, population)
}
\description{
Convert stochastic realisation of branching process into number of new infections time-series
}
\seealso{
Other post-processing: 
\code{\link{generate_hospitalisation_time_series}()},
\code{\link{generate_number_shedding_time_series}()},
\code{\link{generate_symptom_onset_time_series}()}
}
\concept{post-processing}
