% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wastewater_surveillance.R
\name{generate_number_shedding_time_series}
\alias{generate_number_shedding_time_series}
\title{Convert stochastic realisation of branching process into number shedding time-series}
\usage{
generate_number_shedding_time_series(
  branching_process_output,
  shedding_dist,
  shedding_relative_SC2,
  method,
  duration
)
}
\arguments{
\item{branching_process_output}{Output from simulate_branching_process}

\item{shedding_dist}{The distribution of amount shed over time following infection, normalised so the amount of shedding on the day
with the maximum amount is 1.}

\item{shedding_relative_SC2}{Amount of shedding that occurs relative to SARS-CoV-2 (required given our pinning to the SC2 data from Hewitt et al)}

\item{method}{The method to use - either "effective_n_shedders" to use the effective number of shedders based on the shedding profile "shedding_dist" (equivalent to Hewitt Model 3) or "n_shedders" to use the number of shedders (equivalent to Hewitt Model 2) on any one day}

\item{duration}{duration of shedding, if method = "n_shedders". Should be NA if using shedding_dist}
}
\description{
This function converts a branching process output into time-series of number shedding,
weighted by the shedding distribution.
}
\seealso{
Other post-processing: 
\code{\link{extract_outbreak_characteristics_time}()},
\code{\link{generate_healthcare_seeking_time_series}()},
\code{\link{generate_infections_time_series}()},
\code{\link{generate_outbreak_size}()},
\code{\link{generate_seropositivity_timeseries}()},
\code{\link{generate_symptom_onset_time_series}()}
}
\concept{post-processing}
