% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_branching_process.R
\name{generate_seropositivity_timeseries}
\alias{generate_seropositivity_timeseries}
\title{Convert stochastic realisation of branching process into hospitalisation time-series}
\usage{
generate_seropositivity_timeseries(branching_process_output, population)
}
\arguments{
\item{branching_process_output}{Output from simulate_branching_process}

\item{population}{The population size used in the simulate_branching_process output}
}
\description{
This function converts a branching process output into time-series of hospitalisations
and hospital occupancy.
}
\seealso{
Other post-processing: 
\code{\link{generate_healthcare_seeking_time_series}()},
\code{\link{generate_infections_time_series}()},
\code{\link{generate_number_shedding_time_series}()},
\code{\link{generate_outbreak_size}()},
\code{\link{generate_symptom_onset_time_series}()}
}
\concept{post-processing}
