% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clinical_surveillance.R
\name{generate_symptom_onset_time_series}
\alias{generate_symptom_onset_time_series}
\title{Convert stochastic realisation of branching process into symptom onset time-series
This function converts a branching process output into time-series of symptom onsets}
\usage{
generate_symptom_onset_time_series(branching_process_output)
}
\arguments{
\item{branching_process_output}{Output from simulate_branching_process}

\item{population}{The population size used in the simulate_branching_process output}
}
\description{
Convert stochastic realisation of branching process into symptom onset time-series
This function converts a branching process output into time-series of symptom onsets
}
\seealso{
Other post-processing: 
\code{\link{extract_outbreak_characteristics_time}()},
\code{\link{generate_healthcare_seeking_time_series}()},
\code{\link{generate_infections_time_series}()},
\code{\link{generate_number_shedding_time_series}()},
\code{\link{generate_outbreak_size}()},
\code{\link{generate_seropositivity_timeseries}()}
}
\concept{post-processing}
