% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_single_outbreak.R
\name{sim_minimal}
\alias{sim_minimal}
\title{Generate a realisation of a single outbreak from a stochastic
branching process}
\usage{
sim_minimal(
  mn_offspring = 0.9,
  disp_offspring = 1,
  max_gen = Inf,
  index_cases = 1,
  initial_immune = 0
)
}
\arguments{
\item{mn_offspring}{The mean of the offspring distribution (R0)}

\item{disp_offspring}{The overdispersion of the offspring distribution. Must be >= 1. When set to 1, equivalent to a Poisson distribution; >1 is a Negative Binomial distribution.}

\item{max_gen}{The maximum number of generations of transmission to simulate - default is Inf but with mn_offspring <1 transmission dies out eventually.}

\item{index_cases}{The number of initial seeding cases associated with the zoonotic spillover.}

\item{initial_immune}{The proportion of the population initially immune.}
}
\description{
These functions specify a minimal model for simulating a single outbreak
of an infectious disease following a spillover event.
}
\details{
This version of the outbreak simulater is designed to run quickly and simply
and cannot accommodate dependence of infectees characteristics on infectors
characteristics (e.g. mutations --> changes in R0, or decreases in R0 based on
decreases in the proportion of the population susceptible over time)

This function simulates a branching process
}
\seealso{
Other simulation: 
\code{\link{sim_single_outbreak}()},
\code{\link{simulate_branching_process}()},
\code{\link{spillover}()}
}
\concept{simulation}
