% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_single_outbreak.R
\name{sim_single_outbreak}
\alias{sim_single_outbreak}
\title{This function takes the branching process output of sim_minimal and creates
a linelist consisting of a dataframe with 1 row per infected individual with
assigned clinical characteristics based on specified probability distributions.}
\usage{
sim_single_outbreak(
  spillover_day,
  index_case_ID,
  mn_offspring = 0.9,
  disp_offspring = 1,
  max_gen = Inf,
  index_cases = 1,
  generation_time_dist = function(n) {
     rgamma(n, shape = 12, rate = 2)
 },
  prob_symptomatic = 0.8,
  infection_to_onset_dist = function(n) {
     rgamma(n, shape = 6, rate = 2)
 },
  prob_severe = 0.3,
  prob_seek_healthcare_non_severe = 0.5,
  prob_seek_healthcare_severe = 0.95,
  onset_to_healthcare_dist = function(n) {
     rgamma(n, shape = 6, rate = 2)
 },
  prob_diagnosis = 0.8,
  healthcare_to_diagnosis_dist = function(n) {
     rgamma(n, shape = 6, rate = 2)
 },
  initial_immune = 0,
  ...
)
}
\arguments{
\item{spillover_day}{time that spillover occurred (taken from output of spillover function)}

\item{index_case_ID}{ID of the index case (the spillover)}

\item{mn_offspring}{The mean of the offspring distribution (R0)}

\item{disp_offspring}{The overdisperion of the offspring distribution. Must be >= 1. When set to 1, equivalent to a Poisson distribution; >1 is a Negative Binomial distribution.}

\item{max_gen}{The maximum number of generations of transmission to simulate - default is Inf but with mn_offspring <1 transmission dies out eventually.}

\item{index_cases}{The number of initial seeding cases associated with the zoonotic spillover.}

\item{generation_time_dist}{The generation time distribution}

\item{prob_symptomatic}{The probability that an infected individual develops symptoms}

\item{infection_to_onset_dist}{The infection to symptom onset distribution}

\item{prob_severe}{The probability that an infection develops severe disease. Note this IS CONDITIONAL on being symptomatic.}

\item{prob_seek_healthcare_non_severe}{The probability that a non-severe infection seeks healthcare.}

\item{prob_seek_healthcare_severe}{The probability that a severe infection seeks healthcare.}

\item{onset_to_healthcare_dist}{The symptom onset to seeking healthcare delay distribution.}

\item{prob_diagnosis}{The probability of diagnosis given healthcare sought. Note this is not modeled as being dependent on severity, beyond the difference in probability of seeking care.}

\item{healthcare_to_diagnosis_dist}{The distribution of time from healthcare seeking to diagnosis}

\item{initial_immune}{The proportion of the population initially immune.}
}
\description{
This function takes the branching process output of sim_minimal and creates
a linelist consisting of a dataframe with 1 row per infected individual with
assigned clinical characteristics based on specified probability distributions.
}
\seealso{
Other simulation: 
\code{\link{sim_minimal}()},
\code{\link{simulate_branching_process}()},
\code{\link{spillover}()}
}
\concept{simulation}
