% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_spillover.R
\name{solve_b}
\alias{solve_b}
\title{Calculate the baseline rate}
\usage{
solve_b(x, d, tmax, P = 365, sigma = 30)
}
\arguments{
\item{x}{Desired mean spillover rate.}

\item{d}{Seasonal forcing amplitude.}

\item{tmax}{Day of the year at which seasonal forcing peaks.}

\item{P}{Period of seasonality in days.}

\item{sigma}{Spread of the seasonal peak.}
}
\value{
The baseline rate `b`.
}
\description{
Calculates the baseline rate `b` required to achieve a specified
mean spillover rate given the seasonal forcing parameters.
}
