% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_spillover.R
\name{spillover}
\alias{spillover}
\title{Simulating spillovers over time}
\usage{
spillover(
  time,
  specify,
  seasonal_period,
  spillover_rate_pars,
  prevalence_pars,
  contact_rate_pars,
  p_infection
)
}
\arguments{
\item{time}{The time period you would like to run the model over (days)}

\item{specify}{The parameters you would like to specify: either "spillover_rate" to specify spillover rate itself  OR "swiss_cheese" to specify factors contributing to spillover rate (prevalence in animals, contact rate between animals and humans, and probability of infection upon contact)
Used if specify == "spillover_rate"}

\item{spillover_rate_pars}{A list of: "tmax" - the day of the year (1-365) that spillover rate peaks, "b" - the baseline spillover rate, outside of a peak, "d" - the annual seasonal forcing (amplitude increase relative to baseline, default = 0), and "sigma" - the spread of the peak.
Used if specify == "swiss_cheese"}

\item{prevalence_pars}{A list of: "tmax" the day of the year (1-365) that prevalence peaks, "b" - the baseline prevalence of infection in the reservoir, outside of a peak, "d" - the annual seasonal forcing (amplitude increase relative to baseline, default = 0), and "sigma" - the spread of the peak.}

\item{contact_rate_pars}{A list of: "tmax" - the day of the year (1-365) that human-reservoir contact rate peaks, "b" - the baseline daily contact rate between people and the reservoir, "d" - the annual seasonal forcing (amplitude increase relative to baseline, default = 0), and "sigma" - the spread of the peak.}

\item{p_infection}{The probability of a person getting infected given contact with an infected animal. This is assumed to be constant.}
}
\description{
Simulating spillovers over time
}
\seealso{
Other simulation: 
\code{\link{sim_minimal}()},
\code{\link{sim_single_outbreak}()},
\code{\link{simulate_branching_process}()}
}
\concept{simulation}
