% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_branching_process.R
\name{calculate_shedding}
\alias{calculate_shedding}
\title{Utility function to calculate shedding given infection time-series and shedding distribution}
\usage{
calculate_shedding(day, infection_counts, shedding_dist)
}
\description{
Utility function to calculate shedding given infection time-series and shedding distribution
}
\concept{utils}
