% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wastewater_surveillance.R
\name{generate_number_shedding_time_series}
\alias{generate_number_shedding_time_series}
\title{Convert stochastic realisation of branching process into number shedding time-series}
\usage{
generate_number_shedding_time_series(
  branching_process_output,
  shedding_dist,
  shedding_relative_SC2
)
}
\arguments{
\item{branching_process_output}{Output from simulate_branching_process}

\item{shedding_dist}{The distribution of amount shed over time following infection, normalised so the amount of shedding on the day
with the maximum amount is 1.}

\item{shedding_relative_SC2}{Amount of shedding that occurs relative to SARS-CoV-2 (required given our pinning to the SC2 data from Hewitt et al)}
}
\description{
This function converts a branching process output into time-series of number shedding,
weighted by the shedding distribution.
}
\seealso{
Other post-processing: 
\code{\link{generate_healthcare_seeking_time_series}()},
\code{\link{generate_infections_time_series}()},
\code{\link{generate_outbreak_size}()},
\code{\link{generate_seropositivity_timeseries}()},
\code{\link{generate_symptom_onset_time_series}()}
}
\concept{post-processing}
