% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_branching_process.R
\name{generate_number_shedding_time_series}
\alias{generate_number_shedding_time_series}
\title{Convert stochastic realisation of branching process into number shedding time-series}
\usage{
generate_number_shedding_time_series(
  branching_process_output,
  shedding_dist,
  population
)
}
\description{
This function converts a branching process output into time-series of number shedding,
weighted by the shedding distribution.
}
\seealso{
Other post-processing: 
\code{\link{generate_hospitalisation_time_series}()},
\code{\link{generate_symptom_onset_time_series}()}
}
\concept{post-processing}
