% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_branching_process.R
\name{simulate_branching_process}
\alias{simulate_branching_process}
\title{Generate a stochastic realisation from a stochastic branching process}
\usage{
simulate_branching_process(
  mn_offspring = 3,
  disp_offspring = 1.1,
  generation_time_dist = function(n) {
     rgamma(n, shape = 12, rate = 2)
 },
  prob_symptomatic = 0.8,
  infection_to_onset_dist = function(n) {
     rgamma(n, shape = 6, rate = 2)
 },
  prob_hosp = 0.35,
  prob_seek_healthcare = 0.9,
  hospitalisation_delay_dist = function(n) {
     rgamma(n, shape = 24, rate = 2)
 },
  hospitalisation_duration_dist = function(n) {
     rgamma(n, shape = 12, rate = 2)
 },
  initial_immune = 0,
  seeding_cases = 10,
  t0 = 0,
  tf = Inf,
  population = 1e+05,
  check_final_size = 10000
)
}
\description{
This function creates a named list containing the output of simulating a stochastic branching
process.
}
\concept{simulation}
