% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_branching_process.R
\name{simulate_branching_process}
\alias{simulate_branching_process}
\title{Generate a stochastic realisation from a stochastic branching process}
\usage{
simulate_branching_process(
  initial_mn_offspring = 0.9,
  disp_offspring = 0.9,
  generation_time_dist = function(n) {
     rgamma(n, shape = 12, rate = 2)
 },
  prob_symptomatic = 0.8,
  infection_to_onset_dist = function(n) {
     rgamma(n, shape = 6, rate = 2)
 },
  prob_severe = 0.35,
  prob_seek_healthcare_non_severe = 0.9,
  prob_seek_healthcare_severe = 0.9,
  onset_to_healthcare_dist = function(n) {
     rgamma(n, shape = 6, rate = 2)
 },
  prob_seroconvert_asymptomatic = 0.8,
  prob_seroconvert_severe = 0.8,
  prob_seroconvert_non_severe = 0.8,
  infection_to_seroconversion_dist = function(n) {
     rgamma(n, shape = 56, rate = 2)

    },
  seroconversion_to_seroreversion_dist = function(n) {
     rgamma(n, shape = 730, rate =
    2)
 },
  prob_beneficial_mutation = 0.75,
  beneficial_mutation_effect_dist = function(n) {
     rexp(n, rate = 10)
 },
  max_mn_offspring = 4,
  annual_spillover_rate = 2,
  spillover_seeding_cases_dist = function() {
     rpois(1, lambda = 5)
 },
  t0 = 0,
  tf = Inf,
  population = 1e+05,
  initial_immune = 0,
  check_final_size = 10000,
  max_num_outbreaks = 10
)
}
\description{
This function creates a named list containing the output of simulating a stochastic branching
process.
}
\concept{simulation}
